refactor: extract std::stacktrace related symbols into mimic++_ext/stacktrace/std-stacktrace.hpp
